+++ src/utils/png.cpp
@@ -10,13 +10,11 @@
 
 #include <png.h>
 #include <cstdio>
+#include <iostream>
 
 #include "global.h"
 #include "utils/pixel.h"
 #include "utils/png.h"
-
-#include "Console.h"
-#define pngPrint getConsole().print
 
 int pngCheck(const char *filename) {
     png_byte header[8];
@@ -26,7 +24,7 @@ int pngCheck(const char *filename) {
 
     FILE *fp = fopen(filename, "rb");
     if (fp == NULL) {
-        pngPrint("Could not open %s", filename);
+        std::cout << "Could not open " << filename << std::endl;
         return -1;
     }
 
@@ -34,7 +32,7 @@ int pngCheck(const char *filename) {
     fclose(fp);
 
     if (png_sig_cmp(header, 0, 8)) {
-        pngPrint("File %s is not a PNG.", filename);
+        std::cout << "File " << filename << " is not a PNG." << std::endl;
         return -2;
     }
 
@@ -55,28 +53,28 @@ int pngLoad(const char *filename, unsigned char **image,
 
     FILE *fp = fopen(filename, "rb");
     if (fp == NULL) {
-        pngPrint("Could not open %s", filename);
+        std::cout << "Could not open " << filename << std::endl;
         return -1;
     }
 
     fread(header, 1, 8, fp);
 
     if (png_sig_cmp(header, 0, 8)) {
-        pngPrint("File %s is not a PNG.", filename);
+        std::cout << "File " << filename << " is not a PNG." << std::endl;
         fclose(fp);
         return -2;
     }
 
     png_structp png_ptr = png_create_read_struct(PNG_LIBPNG_VER_STRING, NULL, NULL, NULL);
     if (!png_ptr) {
-        pngPrint("png_create_read_struct returned 0.");
+        std::cout << "png_create_read_struct returned 0." << std::endl;
         fclose(fp);
         return -3;
     }
 
     png_infop info_ptr = png_create_info_struct(png_ptr);
     if (!info_ptr) {
-        pngPrint("png_create_info_struct returned 0.");
+        std::cout << "png_create_info_struct returned 0." << std::endl;
         png_destroy_read_struct(&png_ptr, (png_infopp)NULL, (png_infopp)NULL);
         fclose(fp);
         return -4;
@@ -84,14 +82,14 @@ int pngLoad(const char *filename, unsigned char **image,
 
     png_infop end_info = png_create_info_struct(png_ptr);
     if (!end_info) {
-        pngPrint("png_create_info_struct returned 0.");
+        std::cout << "png_create_info_struct returned 0." << std::endl;
         png_destroy_read_struct(&png_ptr, &info_ptr, (png_infopp) NULL);
         fclose(fp);
         return -5;
     }
 
     if (setjmp(png_jmpbuf(png_ptr))) {
-        pngPrint("Error from libpng");
+        std::cout << "Error from libpng" << std::endl;
         png_destroy_read_struct(&png_ptr, &info_ptr, &end_info);
         fclose(fp);
         return -6;
@@ -111,7 +109,7 @@ int pngLoad(const char *filename, unsigned char **image,
     *height = tmpHeight;
 
     if (bit_depth != 8) {
-        pngPrint("%s: Unsupported bit depth %d.  Must be 8.", filename, bit_depth);
+        std::cout << filename << ": Unsupported bit depth " << bit_depth << ".  Must be 8." << std::endl;
         return -7;
     }
 
@@ -141,7 +139,7 @@ int pngLoad(const char *filename, unsigned char **image,
         *mode = RGBA;
         *bpp = 32;
     } else {
-        pngPrint("%s: Unknown libpng color type %d.", filename, color_type);
+        std::cout << filename << ": Unknown libpng color type " << color_type << std::endl;
         delete [] image_data;
         return -8;
     }
@@ -175,27 +173,27 @@ int pngSave(const char *filename, unsigned char *image,
 
     FILE *fp = fopen(filename, "wb");
     if (!fp) {
-        pngPrint("File %s could not be opened for writing", filename);
+        std::cout << "File " << filename << " could not be opened for writing" << std::endl;
         return -1;
     }
 
     png_structp png_ptr = png_create_write_struct(PNG_LIBPNG_VER_STRING, NULL, NULL, NULL);
 
     if (!png_ptr) {
-        pngPrint("png_create_write_struct failed");
+        std::cout << "png_create_write_struct failed" << std::endl;
         fclose(fp);
         return -2;
     }
 
     png_infop info_ptr = png_create_info_struct(png_ptr);
     if (!info_ptr) {
-        pngPrint("png_create_info_struct failed");
+        std::cout << "png_create_info_struct failed" << std::endl;
         fclose(fp);
         return -3;
     }
 
     if (setjmp(png_jmpbuf(png_ptr))) {
-        pngPrint("Error during init_io");
+        std::cout << "Error during init_io" << std::endl;
         fclose(fp);
         return -4;
     }
@@ -203,7 +201,7 @@ int pngSave(const char *filename, unsigned char *image,
     png_init_io(png_ptr, fp);
 
     if (setjmp(png_jmpbuf(png_ptr))) {
-        pngPrint("Error during writing header");
+        std::cout << "Error during writing header" << std::endl;
         fclose(fp);
         return -5;
     }
@@ -222,7 +220,7 @@ int pngSave(const char *filename, unsigned char *image,
         }
         color_type = PNG_COLOR_TYPE_RGB_ALPHA;
     } else {
-        pngPrint("Error invalid color mode");
+        std::cout << "Error invalid color mode" << std::endl;
         fclose(fp);
         return -6;
     }
@@ -239,7 +237,7 @@ int pngSave(const char *filename, unsigned char *image,
         row_pointers[height - 1 - i] = image + (i * width * 4);
 
     if (setjmp(png_jmpbuf(png_ptr))) {
-        pngPrint("Error during writing bytes");
+        std::cout << "Error during writing bytes" << std::endl;
         delete [] row_pointers;
         fclose(fp);
         return -7;
@@ -248,7 +246,7 @@ int pngSave(const char *filename, unsigned char *image,
     png_write_image(png_ptr, row_pointers);
 
     if (setjmp(png_jmpbuf(png_ptr))) {
-        pngPrint("Error during end of write");
+        std::cout << "Error during end of write" << std::endl;
         delete [] row_pointers;
         fclose(fp);
         return -8;